FIX-1937 refactor ContentTable 2
