+++ Src/WitsmlExplorer.Frontend/components/ContentViews/table/ContentTable.tsx
@@ -1,10 +1,10 @@
-import { Checkbox, IconButton, TableBody, TableCell, TableHead, useTheme } from "@material-ui/core";
+import { TableBody, TableHead, useTheme } from "@material-ui/core";
 import {
-  ColumnDef,
   Header,
   Row,
+  RowData,
+  RowSelectionState,
   SortDirection,
-  SortingFns,
   Table,
   flexRender,
   getCoreRowModel,
@@ -14,35 +14,23 @@ import {
 } from "@tanstack/react-table";
 import { useVirtualizer } from "@tanstack/react-virtual";
 import * as React from "react";
-import { Fragment, useEffect, useMemo, useState } from "react";
-import styled from "styled-components";
+import { Fragment, useState } from "react";
 import { indexToNumber } from "../../../models/logObject";
-import { colors } from "../../../styles/Colors";
 import Icon from "../../../styles/Icons";
+import { useColumnDef } from "./ColumnDef";
 import Panel from "./Panel";
-import {
-  ContentTableColumn,
-  ContentTableProps,
-  ContentType,
-  activeId,
-  calculateColumnWidth,
-  constantTableOptions,
-  expanderId,
-  getFromStorage,
-  hiddenStorageKey,
-  orderingStorageKey,
-  saveToStorage,
-  selectId,
-  widthsStorageKey
-} from "./tableParts";
+import { initializeColumnVisibility, useStoreVisibilityEffect, useStoreWidthsEffect } from "./contentTableStorage";
+import { StyledResizer, StyledTable, StyledTd, StyledTh, StyledTr, TableContainer } from "./contentTableStyles";
+import { calculateRowHeight, constantTableOptions, expanderId, isClickable, measureSortingFn, selectId, toggleRow, useInitActiveCurveFiltering } from "./contentTableUtils";
+import { ContentTableColumn, ContentTableProps } from "./tableParts";
 
-const initializeColumnVisibility = (viewId: string | null) => {
-  if (viewId == null) {
-    return {};
+declare module "@tanstack/react-table" {
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  interface TableMeta<TData extends RowData> {
+    previousIndex: number;
+    setPreviousIndex: (index: number) => void;
   }
-  const hiddenColumns = getFromStorage(viewId, hiddenStorageKey);
-  return hiddenColumns == null ? {} : Object.assign({}, ...hiddenColumns.map((hiddenColumn) => ({ [hiddenColumn]: false })));
-};
+}
 
 export const ContentTable = (contentTableProps: ContentTableProps): React.ReactElement => {
   const {
@@ -51,7 +39,6 @@ export const ContentTable = (contentTableProps: ContentTableProps): React.ReactE
     onSelect,
     onContextMenu,
     checkableRows,
-    onRowSelectionChange,
     insetColumns,
     panelElements,
     showPanel = true,
@@ -59,154 +46,14 @@ export const ContentTable = (contentTableProps: ContentTableProps): React.ReactE
     stickyLeftColumns = false,
     viewId
   } = contentTableProps;
-  const [activeIndex, setActiveIndex] = useState<number>(null);
-  const [rowSelection, setRowSelection] = React.useState({});
-  const [columnVisibility, setColumnVisibility] = React.useState(initializeColumnVisibility(viewId));
+  const [previousIndex, setPreviousIndex] = useState<number>(null);
+  const [rowSelection, setRowSelection] = useState<RowSelectionState>({});
+  const [columnVisibility, setColumnVisibility] = useState(initializeColumnVisibility(viewId));
   const isCompactMode = useTheme().props.MuiCheckbox?.size === "small";
   const cellHeight = isCompactMode ? 30 : 53;
   const headCellHeight = isCompactMode ? 35 : 55;
 
-  const selectRow = (e: React.MouseEvent<HTMLButtonElement | HTMLDivElement, MouseEvent>, currentRow: Row<any>, table: Table<any>) => {
-    if (onSelect) {
-      onSelect(currentRow.original);
-    } else if (checkableRows) {
-      toggleRow(e, currentRow, table);
-    }
-  };
-
-  const toggleRow = (e: React.MouseEvent<HTMLButtonElement | HTMLDivElement, MouseEvent>, currentRow: Row<any>, table: Table<any>) => {
-    if (e.shiftKey && activeIndex != null) {
-      const fromIndex = Math.min(activeIndex, currentRow.index);
-      const toIndex = Math.max(activeIndex, currentRow.index);
-      const rows = table.getSortedRowModel().rows;
-      const sortedFromIndex = rows.findIndex((row) => fromIndex == row.index);
-      if (sortedFromIndex == -1) {
-        return;
-      }
-      const newSelections: any = {};
-      let sortedCurrentIndex = sortedFromIndex;
-      while (sortedCurrentIndex < rows.length) {
-        const currentIndex = rows[sortedCurrentIndex].index;
-        newSelections[currentIndex] = true;
-        if (currentIndex == toIndex) {
-          break;
-        }
-        sortedCurrentIndex += 1;
-      }
-      setRowSelection({ ...newSelections, ...rowSelection });
-    } else {
-      currentRow.toggleSelected();
-    }
-    setActiveIndex(currentRow.index);
-  };
-
-  const columnDef = useMemo(() => {
-    const savedWidths = getFromStorage(viewId, widthsStorageKey);
-    let columnDef: ColumnDef<any, any>[] = columns.map((column) => {
-      return {
-        id: column.label,
-        accessorKey: column.property,
-        header: column.label,
-        size: savedWidths ? savedWidths[column.label] : calculateColumnWidth(column.label, isCompactMode, column.type),
-        enableColumnFilter: true,
-        meta: { type: column.type },
-        ...(column.label == activeId
-          ? {
-              filterFn: (row) => row.original.isVisibleFunction(),
-              cell: ({ row }) => {
-                return row.original.isActive ? <Icon name="isActive" /> : "";
-              }
-            }
-          : {}),
-        ...(column.type == ContentType.Measure
-          ? {
-              sortingFn: "measure" as keyof SortingFns
-            }
-          : {})
-      };
-    });
-    const savedOrder = getFromStorage(viewId, orderingStorageKey);
-    //can extract this into a function and write a test
-    if (savedOrder) {
-      const sortedColumns = savedOrder.flatMap((label) => {
-        const foundColumn = columnDef.find((col) => col.id == label);
-        return foundColumn == null ? [] : foundColumn;
-      });
-      const columnsWithoutOrder = columnDef.filter((col) => !savedOrder.includes(col.id));
-      columnDef = sortedColumns.concat(columnsWithoutOrder);
-    }
-
-    if (insetColumns != null) {
-      columnDef = [
-        {
-          id: expanderId,
-          enableHiding: false,
-          size: calculateColumnWidth(expanderId, isCompactMode),
-          header: ({ table }: { table: Table<any> }) => (
-            <IconButton onClick={() => table.toggleAllRowsExpanded(!table.getIsSomeRowsExpanded())} size="small" style={{ padding: 0 }}>
-              <Icon name={table.getIsSomeRowsExpanded() ? "chevronUp" : "chevronDown"} />
-            </IconButton>
-          ),
-          cell: ({ row }) => {
-            return row.getCanExpand() ? (
-              <div style={{ display: "flex" }}>
-                <IconButton
-                  onClick={(event) => {
-                    event.stopPropagation();
-                    row.getToggleExpandedHandler()();
-                  }}
-                  size="small"
-                  style={{ margin: "auto", padding: 0 }}
-                >
-                  <Icon name={row.getIsExpanded() ? "chevronUp" : "chevronDown"} />
-                </IconButton>
-              </div>
-            ) : (
-              ""
-            );
-          }
-        },
-        ...columnDef
-      ];
-    }
-    if (checkableRows) {
-      columnDef = [
-        {
-          id: selectId,
-          enableHiding: false,
-          size: calculateColumnWidth(selectId, isCompactMode),
-          header: ({ table }: { table: Table<any> }) => (
-            <Checkbox
-              {...{
-                checked: table.getIsAllRowsSelected(),
-                indeterminate: table.getIsSomeRowsSelected(),
-                onChange: table.getToggleAllRowsSelectedHandler()
-              }}
-            />
-          ),
-          cell: ({ row, table }: { row: Row<any>; table: Table<any> }) => (
-            <div style={{ display: "flex" }}>
-              <Checkbox
-                style={{ margin: "auto" }}
-                {...{
-                  checked: row.getIsSelected(),
-                  disabled: !row.getCanSelect(),
-                  onClick: (event) => {
-                    toggleRow(event, row, table);
-                    event.stopPropagation();
-                  },
-                  readOnly: true
-                }}
-              />
-            </div>
-          )
-        },
-        ...columnDef
-      ];
-    }
-    return columnDef;
-  }, [columns]);
-
+  const columnDef = useColumnDef(viewId, columns, insetColumns, checkableRows);
   const table = useReactTable({
     data: data ?? [],
     columns: columnDef,
@@ -215,7 +62,7 @@ export const ContentTable = (contentTableProps: ContentTableProps): React.ReactE
       columnVisibility
     },
     sortingFns: {
-      measure: (rowA: Row<any>, rowB: Row<any>, columnId: string) => {
+      [measureSortingFn]: (rowA: Row<any>, rowB: Row<any>, columnId: string) => {
         const a = indexToNumber(rowA.getValue(columnId));
         const b = indexToNumber(rowB.getValue(columnId));
         return a > b ? -1 : a < b ? 1 : 0;
@@ -227,28 +74,23 @@ export const ContentTable = (contentTableProps: ContentTableProps): React.ReactE
     getFilteredRowModel: getFilteredRowModel(),
     getRowCanExpand: insetColumns != null ? (row) => !!row.original.inset?.length : undefined,
     onColumnVisibilityChange: setColumnVisibility,
-    onRowSelectionChange: setRowSelection,
+    onRowSelectionChange: (updaterOrValue) => {
+      const newRowSelection = updaterOrValue instanceof Function ? updaterOrValue(rowSelection) : updaterOrValue;
+      setRowSelection(newRowSelection);
+      // call onRowSelectionChange here with original rows filtered on newRowSelection once VirtualizedContentTable is replaced
+    },
+    meta: {
+      previousIndex,
+      setPreviousIndex
+    },
     enableExpanding: insetColumns != null,
     enableRowSelection: checkableRows,
     ...constantTableOptions
   });
 
-  useEffect(() => {
-    if (onRowSelectionChange) {
-      onRowSelectionChange(
-        table.getSelectedRowModel().rows.map((row) => row.original),
-        null,
-        null
-      );
-    }
-  }, [rowSelection]);
-
-  useEffect(() => {
-    if (viewId != null) {
-      const hiddenColumns = Object.entries(columnVisibility).flatMap(([columnId, isVisible]) => (isVisible ? [] : columnId));
-      saveToStorage(viewId, hiddenStorageKey, hiddenColumns);
-    }
-  }, [columnVisibility]);
+  useStoreWidthsEffect(viewId, table);
+  useStoreVisibilityEffect(viewId, columnVisibility);
+  useInitActiveCurveFiltering(table);
 
   const tableContainerRef = React.useRef<HTMLDivElement>(null);
   const { rows } = table.getRowModel();
@@ -259,20 +101,14 @@ export const ContentTable = (contentTableProps: ContentTableProps): React.ReactE
     estimateSize: () => cellHeight
   });
 
-  useEffect(() => {
-    //use debounce
-    if (viewId != null) {
-      const widths = Object.assign({}, ...table.getLeafHeaders().map((header) => ({ [header.id]: header.getSize() })));
-      saveToStorage(viewId, widthsStorageKey, widths);
+  const onHeaderClick = (e: React.MouseEvent<HTMLDivElement, MouseEvent>, header: Header<any, unknown>) => {
+    if (header.column.getCanSort()) {
+      header.column.getToggleSortingHandler()(e);
+      //collapse all rows to avoid wrong height calculations when sorting expanded rows
+      table.toggleAllRowsExpanded(false);
+      rowVirtualizer.measure();
     }
-  }, [table.getTotalSize()]);
-
-  useEffect(() => {
-    table
-      .getVisibleLeafColumns()
-      .find((col) => col.columnDef.id == activeId)
-      ?.setFilterValue(false);
-  }, [table]);
+  };
 
   const onRowContextMenu = (e: React.MouseEvent<HTMLTableRowElement, MouseEvent>, row: Row<any>) => {
     if (onContextMenu) {
@@ -284,12 +120,11 @@ export const ContentTable = (contentTableProps: ContentTableProps): React.ReactE
     }
   };
 
-  const onHeaderClick = (e: React.MouseEvent<HTMLDivElement, MouseEvent>, header: Header<any, unknown>) => {
-    if (header.column.getCanSort()) {
-      header.column.getToggleSortingHandler()(e);
-      //collapse all rows to avoid wrong height calculations when sorting expanded rows
-      table.toggleAllRowsExpanded(false);
-      rowVirtualizer.measure();
+  const onSelectRow = (e: React.MouseEvent<HTMLButtonElement | HTMLDivElement, MouseEvent>, currentRow: Row<any>, table: Table<any>) => {
+    if (onSelect) {
+      onSelect(currentRow.original);
+    } else if (checkableRows) {
+      toggleRow(e, currentRow, table);
     }
   };
 
@@ -351,8 +186,8 @@ export const ContentTable = (contentTableProps: ContentTableProps): React.ReactE
                       return (
                         <StyledTd
                           key={cell.id}
-                          style={{ width: cell.column.getSize(), left: cell.column.getStart() }}
-                          onClick={clickable ? (event) => selectRow(event, row, table) : undefined}
+                          style={{ width: cell.column.getSize(), left: cell.column.getStart(), height: cellHeight }}
+                          onClick={clickable ? (event) => onSelectRow(event, row, table) : undefined}
                           clickable={clickable ? 1 : 0}
                           sticky={stickyLeftColumns ? 1 : 0}
                         >
@@ -378,17 +213,6 @@ const sortingIcons = {
   asc: <Icon size={16} name="arrowUp" style={{ position: "relative", top: 3 }} />,
   desc: <Icon size={16} name="arrowDown" style={{ position: "relative", top: 3 }} />
 };
-
-function isClickable(onSelect: any, id: string, checkableRows: boolean): boolean {
-  return (onSelect != null || checkableRows) && id != selectId && id != expanderId;
-}
-
-function calculateRowHeight(row: Row<any>, headCellHeight: number, cellHeight: number): number {
-  if (row.getIsExpanded() && row.original.inset?.length != 0) {
-    return headCellHeight + cellHeight + cellHeight * row.original.inset?.length ?? 0;
-  }
-  return cellHeight;
-}
 
 const Resizer = (props: { header: Header<any, unknown> }): React.ReactElement => {
   const { header } = props;
@@ -441,95 +265,3 @@ const Inset = (props: InsetProps): React.ReactElement => {
     </tr>
   );
 };
-
-const TableContainer = styled.div<{ showPanel?: boolean }>`
-  overflow-y: auto;
-  height: 100%;
-  ${(props) =>
-    props.showPanel
-      ? `
-    display: grid;
-    grid-template-rows: 50px 1fr;
-  `
-      : ""}
-`;
-
-const StyledTable = styled.table`
-  width: 100%;
-  border-spacing: 0;
-`;
-
-const StyledResizer = styled.div<{ isResizing?: boolean }>`
-  right: 0;
-  top: 0;
-  position: absolute;
-  height: 100%;
-  width: 7px;
-  background: rgba(0, 0, 0, 0.5);
-  cursor: col-resize;
-  user-select: none;
-  touch-action: none;
-  opacity: 0;
-  ${(props) =>
-    props.isResizing
-      ? `&{
-    background: ${colors.infographic.primaryMossGreen};
-    opacity: 1;
-  }`
-      : ""}
-  &:hover {
-    opacity: 1;
-  }
-`;
-
-const StyledTr = styled.tr<{ selected?: boolean }>`
-  display: flex;
-  width: fit-content;
-  position: absolute;
-  top: 0;
-  left: 0;
-  &&& {
-    background-color: ${(props) => (props.selected ? colors.interactive.textHighlight : "white")};
-  }
-  &&&:nth-of-type(even) {
-    background-color: ${(props) => (props.selected ? colors.interactive.textHighlight : colors.interactive.tableHeaderFillResting)};
-  }
-  &&&:hover {
-    background-color: ${colors.interactive.tableCellFillActivated};
-  }
-`;
-
-const StyledTh = styled(TableCell)<{ sticky?: number }>`
-  && {
-    border-right: 1px solid rgba(224, 224, 224, 1);
-    border-bottom-width: 2px;
-    background-color: ${colors.interactive.tableHeaderFillResting};
-    color: ${colors.text.staticIconsDefault};
-    text-align: center;
-    font-family: EquinorMedium, Arial, sans-serif;
-    position: relative;
-  }
-  > div {
-    font-feature-settings: "tnum";
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-  ${(props) => (props.sticky ? "&:nth-child(1) { position: sticky; z-index: 3; } &:nth-child(2) { position: sticky; z-index: 3; }" : "")}
-`;
-
-const StyledTd = styled(TableCell)<{ clickable?: number; sticky?: number }>`
-  border-right: 1px solid rgba(224, 224, 224, 1);
-  background-color: inherit;
-  z-index: 0;
-  && {
-    color: ${colors.text.staticIconsDefault};
-    font-family: EquinorMedium;
-  }
-  cursor: ${(props) => (props.clickable ? "pointer" : "arrow")};
-  font-feature-settings: "tnum";
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  ${(props) => (props.sticky ? "&:nth-child(1) { position: sticky; z-index: 2; } &:nth-child(2) { position: sticky; z-index: 2; }" : "")}
-`;
